Refactor creation of default files, fixes #581
